requirements.txt
streamlit>=1.39.0
pandas>=2.2.0
numpy>=2.0.0
tensorflow>=2.18.0
scikit-learn>=1.5.0
matplotlib>=3.9.0
sqlalchemy>=2.0.35
psycopg2-binary>=2.9.9
python-dotenv>=1.1.0 